+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Unreleased
-* [#48] Add possibility to raise errors if there are warnings or errors in attribute form when adding feature
-* [#46] Use session scope in qgis_bot fixture
-* [#45] Clean map layers automatically
+* [#47](https://github.com/GispoCoding/pytest-qgis/pull/48) Remove deprecated functionality:
+  * new_project fixture
+  * module_qgis_bot fixture
+  * clean_qgis_layer function
+* [#48](https://github.com/GispoCoding/pytest-qgis/pull/48) Add possibility to raise errors if there are warnings or errors in attribute form when adding feature
+* [#46](https://github.com/GispoCoding/pytest-qgis/pull/46) Use session scope in qgis_bot fixture
+* [#45](https://github.com/GispoCoding/pytest-qgis/pull/45) Clean map layers automatically
 
 # Version 1.3.5 (30-06-2023)
 * [#34](https://github.com/GispoCoding/pytest-qgis/pull/34) Use tempfile instead of protected TempPathFactory in QGIS config path creation